Feature backend dongnp

+++ back-end/AssetManagementTeam6.API/Controllers/AssetController.cs
@@ -2,9 +2,7 @@ using AssetManagementTeam6.API.Attributes;
 using AssetManagementTeam6.API.Dtos.Pagination;
 using AssetManagementTeam6.API.Dtos.Requests;
 using AssetManagementTeam6.API.Services.Interfaces;
-using AssetManagementTeam6.Data.Entities;
 using Common.Enums;
-using FluentValidation.AspNetCore;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -17,10 +15,12 @@ namespace AssetManagementTeam6.API.Controllers
     {
         private readonly IAssetService _assetService;
         private readonly IUserService _userService;
-        public AssetController(IAssetService assetService,IUserService userService)
+        private readonly IAssignmentService _assignmentService;
+        public AssetController(IAssetService assetService,IUserService userService, IAssignmentService assignmentService)
         {
             _assetService = assetService;
-            _userService = userService; 
+            _userService = userService;
+            _assignmentService = assignmentService;
         }
 
         [HttpGet()]
@@ -108,6 +108,13 @@ namespace AssetManagementTeam6.API.Controllers
             if (asset == null)
                 return StatusCode(500, "Can't found asset in the system");
 
+            var assignedAsset = await _assignmentService.GetAssignmentByAssignedAsset(id);
+
+            if (assignedAsset != null)
+            {
+                return StatusCode(500, "Cannot delete asset because it belongs to one or more historical assignments");
+            }
+                
             await _assetService.Delete(id);
 
             return Ok(asset);
